pyproject.toml
[project]
name = "pyrecord"
version = "0.0.1"
description = "A client library for iRecord similar to pyinaturalist"
readme = "README.md"
requires-python = ">=3.10"
dependencies = ["httpx", "python-dotenv", "requests"]

[project.optional-dependencies]
test = ["pytest"]
lint = ["ruff"]
dev = ["pyrecord[test,lint]"]


